requirements.txt
pysondb
pyrogram
tgcrypto
pyromod
apscheduler
geopy
pydantic
python-dotenv